🚸 Expose `.namachine` Export to Python to interface with MQT NAViz (#719)
## Description

The C++ library already had a function that facilitated the export of an
architecture specification to the `.namachine` format used by MQT NAViz.
This PR eposes this functionality to Python now via two functions:
Either the specification is returned as a string or written to a file.

## Checklist:

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: Yannick Stade <100073938+ystade@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>